Decalist_Codebase/requirements.txt
beautifulsoup4==4.12.3
Django==5.0.2
djangorestframework==3.14.0
duckduckgo_search==7.1.1
httpx==0.28.1
jmespath==1.0.1
protobuf==3.20.0
python-dotenv==1.0.1
Requests==2.32.3
selenium==4.27.1
